backend/auth/requirements.txt
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic[email]==2.5.0
bcrypt==4.1.2
PyJWT==2.8.0